+++ CHANGES.md
@@ -1,2 +1,3 @@
 24 06 23 :
 - parallelization, 1 neutron per process, several positions (one trajectory) per neutron.
+- starting homemade serialization to pass object Trajectory.